+++ app/controllers/admin/users_controller.rb
@@ -1,10 +1,11 @@
 class Admin::UsersController < Admin::BaseController
-  before_action :set_tenant
+  before_action :current_tenant
   before_action :set_user, only: %i[show edit update destroy]
 
   # GET /users or /users.json
   def index
     @users = @tenant.users.all
+    @users = case_insensitive_match(@users, [ :name, :email, :phone ])
   end
 
   # GET /users/1 or /users/1.json
@@ -18,6 +19,10 @@ class Admin::UsersController < Admin::BaseController
 
   # GET /users/1/edit
   def edit
+    if @user.superuser? && !current_user.superuser?
+      redirect_to admin_tenant_users_path(@tenant), alert: "You cannot edit a superuser."
+      nil
+    end
   end
 
   # POST /users or /users.json
@@ -46,15 +51,20 @@ class Admin::UsersController < Admin::BaseController
     redirect_to admin_tenant_users_path(@tenant), status: :see_other, notice: "User was successfully destroyed."
   end
 
+  def delete_all
+    @users = @tenant.users # current_user.superuser? ? Device.all : Device.where(tenant_id: current_user.tenant_id)
+    case_insensitive_match(@users, [ :name, :email, :phone ]).destroy_all
+
+    respond_to do |format|
+      format.html { redirect_to admin_tenant_users_path(@tenant), status: :see_other, notice: "All users were successfully deleted." }
+      format.json { head :no_content }
+    end
+  end
+
   private
 
   def set_user
     @user = @tenant.users.find(params[:id])
-  end
-
-  def set_tenant
-    @tenant = Tenant.find(params[:tenant_id]) || current_user.tenant
-    redirect_to admin_login_path unless @tenant
   end
 
   def user_params