+++ app/helpers/application_helper.rb
@@ -1,2 +1,5 @@
 module ApplicationHelper
+  def query_param
+    session[:query] || @query || ""
+  end
 end