+++ app/javascript/controllers/search_controller.js
@@ -14,17 +14,26 @@ export default class extends Controller {
     clearTimeout(this.timeout)
   }
 
+  keydown(event) {
+    // If the user presses Enter, we want to submit the form
+    if (event.key === "Enter") {
+      event.preventDefault()
+      this.search(event)
+    }
+  }
+
   search(event) {
     clearTimeout(this.timeout)
 
-    const query = event.target.value.trim()
+    let query = event.target.value.trim()
 
     this.timeout = setTimeout(() => {
-      if (query) {
-        const url = new URL(window.location.href)
-        url.searchParams.set("query", query)
-        window.location.href = url.toString()
+      if (query === "") {
+        query = "*"
       }
+      const url = new URL(window.location.href)
+      url.searchParams.set("query", query)
+      window.location.href = url.toString()
     }, this.delayValue)
   }
 }